requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.39.1
django-filter==2.1.0
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
idna==2.8
oauthlib==3.0.1
psycopg2==2.8.2
psycopg2-binary==2.8.2
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
trello==0.9.3
urllib3==1.25.2
whitenoise==4.1.2
